Refactor prediction, ensemble algorithms out of pipeline.py
+ Prediction algorithms now go in inferrers.py
+ Ensemble algorithms now go in ensembles.py
